wanted to see how A* would behave with a perfect heuristic
that is, a heuristic that literally knows the exact shortest path to the
target (obtained by running A* on each open node (sloth))

super slow, but it does work! guides the sloth straight to the target.

however, there might be an issue with the handling of open nodes since
manhatten distance as heuristic doesn't quite work, not sure if that
affects this though.